feat [DZ]: refactor course model imports and remove obsolete course_model.dart file
