+++ lib/model/course_model.dart
