+++ lib/widget/course_page/note_section.dart
@@ -1,4 +1,4 @@
-import 'package:admin_ocean_learn2/pages/course_page/course_model.dart';
+import 'package:admin_ocean_learn2/model/course_model.dart';
 import 'package:flutter/material.dart';
 
 class NotesSection extends StatelessWidget {
